requirements.txt
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi==2023.7.22
charset-normalizer==3.2.0
coloredlogs==15.0.1
contourpy==1.0.1
cycler==0.11.0
-e git+https://github.com/JunkyByte/easy_ViTPose@66cf64690d03a9d5ddd5bde3a9470cbd8cad752d#egg=easy_ViTPose
ffmpeg==1.4
filelock==3.12.2
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.43.0
fsspec==2024.9.0
humanfriendly==10.0
idna==3.4
imageio==2.31.3
importlib-resources==6.1.0
Jinja2==3.1.4
joblib @ file:///croot/joblib_1718217211762/work
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
onnx==1.14.1
onnxruntime==1.16.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.1
pillow==10.4.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pooch @ file:///croot/pooch_1695850093751/work
protobuf==4.24.3
psutil==5.9.5
py-cpuinfo==9.0.0
pycocotools==2.0.8
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn @ file:///croot/scikit-learn_1721921875708/work
scipy==1.11.2
seaborn==0.12.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.12
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tifffile==2023.9.18
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.1
triton==3.0.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.2.48
ultralytics-thop==2.0.9
urllib3 @ file:///croot/urllib3_1727769808118/work
zipp==3.17.0
